+++ src/main/java/com/project/cheerha/common/config/WebConfig.java
@@ -2,13 +2,11 @@ package com.project.cheerha.common.config;
 
 import java.util.List;
 
-import com.project.cheerha.common.interceptor.IpBlockingInterceptor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.filter.ForwardedHeaderFilter;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @Configuration
@@ -16,7 +14,6 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 public class WebConfig implements WebMvcConfigurer {
 
     private final AuthUserArgumentResolver authUserArgumentResolver;
-    private final IpBlockingInterceptor ipBlockingInterceptor;
 
     @Override
     public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
@@ -26,12 +23,5 @@ public class WebConfig implements WebMvcConfigurer {
     @Bean
     ForwardedHeaderFilter forwardedHeaderFilter() {
         return new ForwardedHeaderFilter();
-    }
-
-    //모든 요청에 ip 밴 적용
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(ipBlockingInterceptor)
-                .addPathPatterns("/**");
     }
 }
